dev: main to 3.3.6

+++ source/libs/sync/inc/syncPipeline.h
@@ -39,7 +39,7 @@ typedef struct SSyncLogReplMgr {
   int64_t       peerStartTime;
   int32_t       retryBackoff;
   int32_t       peerId;
-  int32_t       sendCount;
+  int64_t       sendCount;
   TdThreadMutex mutex;
 } SSyncLogReplMgr;
